+++ R/stacked-bar.R
@@ -101,7 +101,7 @@ draw_stack_barplot <- function(
 
     plot_data <- as.data.frame(tab, stringsAsFactors = FALSE)
     names(plot_data) <- c(".x", ".y", "n")
-    totals <- ave(plot_data$n, plot_data$.x, FUN = sum)
+    totals <- stats::ave(plot_data$n, plot_data$.x, FUN = sum)
     plot_data$prop <- plot_data$n / totals
     plot_data$label <- paste0(
       plot_data$n, " (", scales::percent(plot_data$prop, accuracy = percent_accuracy), ")"